# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/chapters/{id}": {
      "get": {
        "operationId": "ChaptersController_getChapter",
        "summary": "Get chapter details",
        "description": "Retrieve chapter information. Includes user progress if authenticated.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Chapter ID",
            "schema": {
              "example": "012e3456-e78b-90d1-a234-426614174333",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chapter retrieved successfully"
          },
          "404": {
            "description": "Chapter not found"
          }
        },
        "tags": [
          "Chapters"
        ]
      },
      "patch": {
        "operationId": "ChaptersController_updateChapter",
        "summary": "Update chapter",
        "description": "Update chapter details (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Chapter ID",
            "schema": {
              "example": "012e3456-e78b-90d1-a234-426614174333",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChapterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chapter updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Chapter not found"
          }
        },
        "tags": [
          "Chapters"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ChaptersController_deleteChapter",
        "summary": "Delete chapter",
        "description": "Delete a chapter (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Chapter ID",
            "schema": {
              "example": "012e3456-e78b-90d1-a234-426614174333",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chapter deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Chapter not found"
          }
        },
        "tags": [
          "Chapters"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chapters/{id}/progress": {
      "patch": {
        "operationId": "ChaptersController_updateProgress",
        "summary": "Update chapter progress",
        "description": "Update video position for resume functionality",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Chapter ID",
            "schema": {
              "example": "012e3456-e78b-90d1-a234-426614174333",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Progress updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Chapter not found"
          }
        },
        "tags": [
          "Chapters"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/chapters/{id}/complete": {
      "post": {
        "operationId": "ChaptersController_completeChapter",
        "summary": "Complete chapter",
        "description": "Mark a chapter as completed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Chapter ID",
            "schema": {
              "example": "012e3456-e78b-90d1-a234-426614174333",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chapter marked as completed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Chapter not found"
          }
        },
        "tags": [
          "Chapters"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/admin/stats": {
      "get": {
        "operationId": "AdminController_getStats",
        "summary": "Get admin dashboard statistics",
        "description": "Retrieve counts and recent activity for admin dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "stats": {
                      "users": 100,
                      "courses": 30,
                      "projects": 50,
                      "enrollments": 200
                    },
                    "recentUsers": [],
                    "recentCourses": []
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "WKU Software Crew API",
    "description": "RESTful API for WKU Software Crew platform - A comprehensive learning and collaboration platform for software development students",
    "version": "1.0",
    "contact": {
      "name": "WKU Software Crew",
      "url": "https://github.com/wku-crew",
      "email": "contact@wku-crew.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateProgressDto": {
        "type": "object",
        "properties": {
          "lastPosition": {
            "type": "number",
            "description": "Last video position in seconds",
            "example": 120,
            "minimum": 0
          }
        }
      },
      "UpdateChapterDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Chapter title",
            "example": "1. Introduction to React"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug",
            "example": "introduction-to-react"
          },
          "order": {
            "type": "number",
            "description": "Order in the course",
            "example": 1
          },
          "duration": {
            "type": "number",
            "description": "Duration in minutes",
            "example": 15
          },
          "videoUrl": {
            "type": "string",
            "description": "Video URL (YouTube, Vimeo, etc.)"
          },
          "content": {
            "type": "string",
            "description": "Chapter content in Markdown"
          }
        }
      },
...
    }
  }
}